libs/ease_llm/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "ease-llm"
version = "0.0.0"
description = "Cliente LLM + loader YAML. Solo uso local."
authors = [{name = "Tu Nombre"}]
dependencies = [
    "openai>=1.30",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
]
requires-python = ">=3.9"
